requirements.txt
groq>=0.4.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.0
beautifulsoup4>=4.12.0
readability-lxml>=0.8.1
lxml>=5.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
playwright>=1.40.0
msal>=1.28.0
